% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{classificador_turma}
\alias{classificador_turma}
\title{Consolidate pages, decisions, and comprot}
\usage{
classificador_turma(id, path)
}
\arguments{
\item{id}{a list containing one or more lawsuits id's (in numeric format if's possible)}

\item{path}{directory where exists the directory of pdf decisions}
}
\value{
A tibble with aggregated data about sections
}
\description{
Consolidate pages, decisions, and comprot
}
