% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_aux.R
\name{classificador_turma_}
\alias{classificador_turma_}
\title{Get informations about section of a lawsuit}
\usage{
classificador_turma_(id, path)
}
\arguments{
\item{id}{a lawsuit id}

\item{path}{directory where exists the directory of pdf decisions}
}
\value{
A tibble
}
\description{
Get informations about section of a lawsuit
}
