% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_aux.R
\name{consolidate_appeals}
\alias{consolidate_appeals}
\title{Consolidate appeals}
\usage{
consolidate_appeals(data)
}
\arguments{
\item{data}{Data returned by [consolidate_states()]}
}
\value{
A tibble
}
\description{
Consolidate appeals
}
