% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_aux.R
\name{consolidate_chambers}
\alias{consolidate_chambers}
\title{Consolidate chambers}
\usage{
consolidate_chambers(data)
}
\arguments{
\item{data}{Data returned by [consolidate_rapporteurs()]}
}
\value{
A tibble
}
\description{
Consolidate chambers
}
