% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_aux.R
\name{consolidate_decisions}
\alias{consolidate_decisions}
\title{Consolidate decisions}
\usage{
consolidate_decisions(pages, decisions)
}
\arguments{
\item{pages}{`pages` element returned by [parse_decision()]}

\item{decisions}{`decisions` element returned by [parse_decision()]}
}
\value{
A tibble
}
\description{
Consolidate decisions
}
