% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_aux.R
\name{consolidate_origins}
\alias{consolidate_origins}
\title{Consolidate origins}
\usage{
consolidate_origins(data, comprot)
}
\arguments{
\item{data}{Data returned by [consolidate_decisions()]}

\item{comprot}{Table returned by [get_comprot()]}
}
\value{
A tibble
}
\description{
Consolidate origins
}
