% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_aux.R
\name{consolidate_rapporteurs}
\alias{consolidate_rapporteurs}
\title{Consolidate rapporteurs}
\usage{
consolidate_rapporteurs(data)
}
\arguments{
\item{data}{Data returned by [consolidate_taxes()]}
}
\value{
A tibble
}
\description{
Consolidate rapporteurs
}
