% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_aux.R
\name{consolidate_results}
\alias{consolidate_results}
\title{Consolidate results}
\usage{
consolidate_results(data)
}
\arguments{
\item{data}{Data returned by [consolidate_votes()]}
}
\value{
A tibble
}
\description{
Consolidate results
}
