% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_aux.R
\name{consolidate_states}
\alias{consolidate_states}
\title{Consolidate states (UF)}
\usage{
consolidate_states(data)
}
\arguments{
\item{data}{Data returned by [consolidate_origins()]}
}
\value{
A tibble
}
\description{
Consolidate states (UF)
}
