% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_aux.R
\name{consolidate_taxes}
\alias{consolidate_taxes}
\title{Consolidate taxes}
\usage{
consolidate_taxes(data)
}
\arguments{
\item{data}{Data returned by [consolidate_results()]}
}
\value{
A tibble
}
\description{
Consolidate taxes
}
