% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_aux.R
\name{consolidate_votes}
\alias{consolidate_votes}
\title{Consolidate votes}
\usage{
consolidate_votes(data)
}
\arguments{
\item{data}{Data returned by [consolidate_appeals()]}
}
\value{
A tibble
}
\description{
Consolidate votes
}
