% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_decision}
\alias{download_decision}
\title{Download CARF decisions}
\usage{
download_decision(id, path = ".", download_pdfs = TRUE)
}
\arguments{
\item{id}{A character vector with one or more decision numbers}

\item{path}{Directory where to save downloaded files}

\item{download_pdfs}{Whether or not to download the attachments}
}
\value{
A character vetor with the paths to the downloaded files
}
\description{
Download CARF decisions
}
