% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_aux.R
\name{download_decision_}
\alias{download_decision_}
\title{Download CARF decision}
\usage{
download_decision_(id, path, interval, download_pdfs)
}
\arguments{
\item{id}{Character vector with a decision numbers}

\item{path}{Directory where to save file}

\item{interval}{A vector with two dates for beginning and end of query}

\item{download_pdfs}{Whether or not to download the attachments}
}
\value{
A character vetor with the path to the downloaded file
}
\description{
Download CARF decision
}
