% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_decision_period}
\alias{download_decision_period}
\title{Download all CARF decisions from a period}
\usage{

  download_decision_period(date_min = stringr::str_c(stringr::str_pad(lubridate::month(lubridate::today()),
  2, "0", side = "left"), "/", lubridate::year(lubridate::today() -
  lubridate::years(1))),
  date_max = stringr::str_c(stringr::str_pad(lubridate::month(lubridate::today()),
  2, "0", side = "left"), "/", lubridate::year(lubridate::today())),
  path = ".", min_page = 1, max_page = Inf)
}
\arguments{
\item{date_min}{First date of the range. The format need to be like 01/2018}

\item{date_max}{Final date of the range (with same year that date_min). The format need to be like 01/2018}

\item{path}{Directory where to save downloaded files}

\item{min_page}{Page from which to start downloading}

\item{max_page}{Page at which to stop downloading}
}
\value{
A character vetor with the paths to the downloaded files
}
\description{
Download all CARF decisions from a period
}
