% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_decision_year}
\alias{download_decision_year}
\title{Download all CARF decisions from a year}
\usage{
download_decision_year(year, path = ".", min_page = 1, max_page = Inf)
}
\arguments{
\item{year}{Year to get decisions from}

\item{path}{Directory where to save downloaded files}

\item{min_page}{Page from which to start downloading}

\item{max_page}{Page at which to stop downloading}
}
\value{
A character vetor with the paths to the downloaded files
}
\description{
Download all CARF decisions from a year
}
