% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_aux.R
\name{download_decision_year_}
\alias{download_decision_year_}
\title{Download a decision form a page}
\usage{
download_decision_year_(dcis_num, r_page, path, pb)
}
\arguments{
\item{dcis_num}{Number of the decision to download}

\item{r_page}{Response from the page query run in [download_year_page()]}

\item{path}{Path to the page directory where to save decision}

\item{pb}{Progress bar generated by [download_year_carf()]}
}
\value{
A character vector with the path to the downloaded file
}
\description{
Download a decision form a page
}
