% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_lawsuit}
\alias{download_lawsuit}
\title{Download lawsuits from the CARF website}
\usage{
download_lawsuit(id, path = ".", download_pdfs = TRUE)
}
\arguments{
\item{id}{A character vector with one or more lawsuit numbers}

\item{path}{The directory where to save downloaded files}

\item{download_pdfs}{Whether or not to download the attachments}
}
\value{
A character vector with the paths to the downloaded files
}
\description{
Download lawsuits from the CARF website
}
