% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_aux.R
\name{download_lawsuit_}
\alias{download_lawsuit_}
\title{Auxiliary function to download one lawsuit}
\usage{
download_lawsuit_(id, path, interval, download_pdfs)
}
\arguments{
\item{id}{Lawsuit number}

\item{path}{Folder where to save files}

\item{interval}{A vector with two dates for beginning and end of query}

\item{download_pdfs}{Whether or not to download the attachments}
}
\value{
Path to downloaded file
}
\description{
Auxiliary function to download one lawsuit
}
