% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_lawsuit_new}
\alias{download_lawsuit_new}
\title{Remasterized download_lawsuits function}
\usage{
download_lawsuit_new(id, path, download_pdfs = TRUE)
}
\arguments{
\item{id}{A character vector with one or more lawsuit numbers (in numeric format if's possible)}

\item{path}{The directory where to save downloaded files}

\item{download_pdfs}{Whether or not to download the attachments}
}
\description{
Remasterized download_lawsuits function
}
