% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_aux.R
\name{download_lawsuit_v2}
\alias{download_lawsuit_v2}
\title{remasterized auxiliary function to download one lawsuit}
\usage{
download_lawsuit_v2(id, path, download_pdfs = TRUE, pb)
}
\arguments{
\item{id}{Lawsuit number}

\item{path}{Folder where to save files}

\item{download_pdfs}{Whether or not to download the attachments}

\item{interval}{A vector with two dates for beginning and end of query}
}
\description{
remasterized auxiliary function to download one lawsuit
}
