% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_aux.R
\name{download_page_year}
\alias{download_page_year}
\title{Download all lawsuits form a page}
\usage{
download_page_year(page_num, r_search, path, pb)
}
\arguments{
\item{page_num}{Number of the page to download}

\item{r_search}{Response from the search query run in [download_year_carf()]}

\item{path}{Path to the page directory where to save decisions}

\item{pb}{Progress bar generated by [download_year_carf()]}
}
\value{
A character vector with the paths to the downloaded files
}
\description{
Download all lawsuits form a page
}
