% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_pdf}
\alias{download_pdf}
\title{Auxiliary function for downloading attachments}
\usage{
download_pdf(info, path, u_pdf, viewstate)
}
\arguments{
\item{info}{Vector with PDF name and _idcl}

\item{path}{Directory where to save PDF}

\item{u_pdf}{URL where to make the POST}

\item{viewstate}{View state from the last page}
}
\description{
Auxiliary function for downloading attachments
}
