% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whos_who.R
\name{download_whos_who}
\alias{download_whos_who}
\title{Download CARF's "Who's Who" ("Quem e Quem")}
\usage{
download_whos_who(path = ".")
}
\arguments{
\item{path}{Directory where to save PDF}
}
\value{
Character vector with the path to the downloaded file
}
\description{
Download CARF's "Who's Who" ("Quem e Quem")
}
