% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprot.R
\name{get_comprot}
\alias{get_comprot}
\title{Get lawsuits' comunication and protocol data (comprot)}
\usage{
get_comprot(id)
}
\arguments{
\item{id}{Character vector with one or more lawsuit IDs}
}
\value{
A tibble with the following columns:
\itemize{
  \item `id` Number of the lawsuit
  \item `date` Date of protocol
  \item `document` ID of the document of origin
  \item `origin` Name of place of origin
  \item `subject` Subject of the lawsuit
  \item `party` Name of the interessed party
  \item `cpf_cnpf` Party's CPF or CNPJ
  \item `status` Lawsuit's status
  \item `state` State where the lawsuit was filed
  \item `events` A tibble with the events the lawsuit's been through
}
}
\description{
Get lawsuits' comunication and protocol data (comprot)
}
