% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprot.R
\name{get_comprot_}
\alias{get_comprot_}
\title{Get the comprot data for one lawsuit}
\usage{
get_comprot_(...)
}
\arguments{
\item{id}{Character vector with lawsuit ID}
}
\value{
One row containing lawsuit's comprot data
}
\description{
Get the comprot data for one lawsuit
}
