% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_viewstate}
\alias{get_viewstate}
\title{Auxiliary function to get viewstate from response}
\usage{
get_viewstate(response)
}
\arguments{
\item{response}{Response from an HTTP request}
}
\description{
Auxiliary function to get viewstate from response
}
