% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jurisprudence_origins}
\alias{jurisprudence_origins}
\title{Information about possible origins for CARF decicions}
\format{A tibble with columns:
\describe{
 \item{id}{First five digits of a lawsuit ID}
 \item{origin}{Name of origin}
}}
\usage{
jurisprudence_origins
}
\description{
Information about possible origins for CARF decicions
}
\keyword{datasets}
