% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_decision}
\alias{parse_decision}
\title{Parse jurisprudence decisions and pages from [download_decision_year()] or
[download_decision()]}
\usage{
parse_decision(file, type = "both", verbose = TRUE)
}
\arguments{
\item{file}{A character vector with the paths to one of more jurisprudence
files (as returned by [download_decision_year()] or [download_decision()])}

\item{type}{Whether to parse for `"page"`, `"decision"` or `"both"`}

\item{verbose}{Whether or not to print a progress bar}
}
\value{
A tibble with the data parsed
}
\description{
Parse jurisprudence decisions and pages from [download_decision_year()] or
[download_decision()]
}
