% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_aux.R
\name{parse_decision_}
\alias{parse_decision_}
\title{Parse jurisprudence decision}
\usage{
parse_decision_(file, verbose, pb)
}
\arguments{
\item{file}{A character vector with the path to a decision}

\item{verbose}{Whether or not to print the progress bar}

\item{pb}{Progress bar created by parse_decision()}
}
\value{
A tibble with one row containing the data parsed
}
\description{
Parse jurisprudence decision
}
