% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_aux.R
\name{parse_div}
\alias{parse_div}
\title{Parse each div from a page}
\usage{
parse_div(div, verbose, pb)
}
\arguments{
\item{div}{Div to parse for relevant decision information}

\item{verbose}{Whether to tick the progress bar}

\item{pb}{Progress bar created by [parse_decision()]}
}
\value{
A tibble with one row of parsed data
}
\description{
Parse each div from a page
}
