% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_lawsuit}
\alias{parse_lawsuit}
\title{Parser for CARF lawsuit files}
\usage{
parse_lawsuit(file)
}
\arguments{
\item{file}{The path to a file returned by [download_lawsuit()]}
}
\value{
A tibble with information about the lawsuit
}
\description{
Parser for CARF lawsuit files
}
