% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_aux.R
\name{puxa_pdf_}
\alias{puxa_pdf_}
\title{Get information about pdf lawsuit}
\usage{
puxa_pdf_(...)
}
\arguments{
\item{path}{directory where exists the directory of pdf decisions}
}
\value{
A list
}
\description{
Get information about pdf lawsuit
}
