% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rapporteur_data}
\alias{rapporteur_data}
\title{Information about rapporteurs}
\format{A tibble with columns:
\describe{
 \item{rapporteur}{Name of the rapporteur}
 \item{title}{Rapporteur's title (position)}
 \item{type}{Rapporteur type}
 \item{chamber}{Chamber where the rapporteur works}
 \item{section}{Section of the chamber}
}}
\usage{
rapporteur_data
}
\description{
Information about rapporteurs
}
\keyword{datasets}
