% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rapporteur_goliva}
\alias{rapporteur_goliva}
\title{Information about rapporteurs from the Goliva dataset}
\format{A tibble with columns:
\describe{
 \item{rapporteur}{Name of the rapporteur}
 \item{n}{Number of times the rapporteur came up}
 \item{title}{Rapporteur's title (position)}
 \item{type}{Rapporteur type}
 \item{chamber}{Chamber where the rapporteur works}
 \item{section}{Section of the chamber}
}}
\usage{
rapporteur_goliva
}
\description{
Information about rapporteurs from the Goliva dataset
}
\keyword{datasets}
