% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whos_who.R
\name{read_whos_who}
\alias{read_whos_who}
\title{Read CARF's "Who's Who" ("Quem e Quem") file}
\usage{
read_whos_who(file)
}
\arguments{
\item{file}{Path to PDF downloaded with [download_whos_who()]}
}
\value{
A tibble with who's who information
}
\description{
Read CARF's "Who's Who" ("Quem e Quem") file
}
