% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whos_who.R
\name{split_len}
\alias{split_len}
\title{Auxiliary function for spliting a string at certain points}
\usage{
split_len(string, splits)
}
\arguments{
\item{string}{String to be split}

\item{splits}{Number of splits to be returned}
}
\value{
A character vector with a split string
}
\description{
Auxiliary function for spliting a string at certain points
}
