% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{year_post}
\alias{year_post}
\title{Helper funcion for making download_year_*()-related POST requests}
\usage{
year_post(url, response, body, file = NULL)
}
\arguments{
\item{url}{Character vector with the request URL}

\item{response}{Last request's response}

\item{body}{Body of the request}

\item{file}{Either `NULL` or the path where to save the response}
}
\value{
The response returned by [httr::POST()]
}
\description{
Helper funcion for making download_year_*()-related POST requests
}
